+++ Packages/MonoHooks/Runtime/Functions/GameObjectGameObjectFunction.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace UnityAtoms.MonoHooks
+{
+    public abstract class GameObjectGameObjectFunction : AtomFunction<GameObject, GameObject> { }
+}