+++ Packages/MonoHooks/Runtime/Hooks/OnDestroyHook.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+
+namespace UnityAtoms.MonoHooks
+{
+    /// <summary>
+    /// Mono Hook for [`OnDestroy`](https://docs.unity3d.com/ScriptReference/MonoBehaviour.OnDestroy.html)
+    /// </summary>
+    [EditorIcon("atom-icon-delicate")]
+    [AddComponentMenu("Unity Atoms/Hooks/On Destroy")]
+    public sealed class OnDestroyHook : VoidHook
+    {
+        private void OnDestroy()
+        {
+            OnHook(new Void());
+        }
+    }
+}