+++ Packages/MonoHooks/Runtime/Hooks/OnUpdateHook.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+
+namespace UnityAtoms.MonoHooks
+{
+    /// <summary>
+    /// Mono Hook for [`Update`](https://docs.unity3d.com/ScriptReference/MonoBehaviour.Update.html)
+    /// </summary>
+    [EditorIcon("atom-icon-delicate")]
+    [AddComponentMenu("Unity Atoms/Hooks/On Update")]
+    public sealed class OnUpdateHook : VoidHook
+    {
+        private void Update()
+        {
+            OnHook(new Void());
+        }
+    }
+}